+++ examples/src/client_pt.rs
@@ -1,9 +1,8 @@
-use examples::{DemoServiceClient, LogAspect, ReqDto};
+use examples::{DemoServiceClient, ReqDto};
 use fusen_rs::fusen_common::config::get_config_by_file;
 use fusen_rs::fusen_common::date_util::get_now_date_time_as_millis;
+use fusen_rs::fusen_common::logs::LogConfig;
 use fusen_rs::fusen_common::register::Type;
-use fusen_rs::handler::HandlerLoad;
-use fusen_rs::register::nacos::NacosConfig;
 use fusen_rs::{fusen_common, FusenApplicationContext};
 use std::sync::Arc;
 use std::time::Duration;
@@ -12,7 +11,13 @@ use tracing::info;
 
 #[tokio::main(flavor = "multi_thread", worker_threads = 2)]
 async fn main() {
-    fusen_common::logs::init_log();
+    let log_config = LogConfig::default()
+    .devmode(Some(true))
+    .env_filter(Some(
+        "fusen-rs=debug,client=debug,examples=debug".to_owned(),
+    ))
+    .endpoint(Some("http://127.0.0.1:4317".to_owned()));
+let _log_work = fusen_common::logs::init_log(&log_config, "fusen-client-pt");
     let context = FusenApplicationContext::builder()
         .init(get_config_by_file("examples/client-config.yaml").unwrap())
         .build();
