+++ examples/src/server.rs
@@ -1,13 +1,13 @@
 use examples::{DemoService, LogAspect, ReqDto, ResDto};
 use fusen_rs::fusen_common::config::get_config_by_file;
 use fusen_rs::fusen_common::date_util::get_now_date_time_as_millis;
-use fusen_rs::fusen_common::logs1::LogConfig;
+use fusen_rs::fusen_common::logs::LogConfig;
 use fusen_rs::fusen_procedural_macro::{asset, handler};
 use fusen_rs::handler::aspect::Aspect;
 use fusen_rs::handler::HandlerLoad;
 use fusen_rs::{fusen_common, FusenApplicationContext};
 use fusen_rs::{fusen_common::FusenResult, fusen_procedural_macro::fusen_server};
-use tracing::info;
+use tracing::{info, info_span};
 
 struct ServerLogAspect;
 
@@ -43,7 +43,12 @@ impl DemoService for DemoServiceImpl {
     }
     #[asset(path="/sayHelloV2-http",method = POST)]
     async fn sayHelloV2(&self, req: ReqDto) -> FusenResult<ResDto> {
-        info!("res : {:?}", req);
+        let _span = info_span!("sayHelloV2-http").entered();
+        info!("开始处理 sayHelloV2-http");
+        info!("接收消息 : {:?}", req);
+        let _span2 = info_span!("get_user_info_by_db").entered();
+        info!("get_user_info_by_db : selet * from user where id = $1");
+        drop(_span2);
         Ok(ResDto::default().str("Hello ".to_owned() + req.get_str() + " V2"))
     }
     #[asset(path="/divide",method = GET)]
@@ -61,7 +66,7 @@ async fn main() {
             "fusen-rs=debug,server=debug,examples=debug".to_owned(),
         ))
         .endpoint(Some("http://127.0.0.1:4317".to_owned()));
-    let _log_work = fusen_common::logs1::init_log(&log_config, "fusen-server");
+    let _log_work = fusen_common::logs::init_log(&log_config, "fusen-server");
     let server = DemoServiceImpl {
         _db: "我是一个DB数据库".to_string(),
     };
