+++ fusen/Cargo.toml
@@ -32,6 +32,10 @@ tracing.workspace = true
 tracing-futures.workspace = true
 tracing-subscriber.workspace = true
 pretty_env_logger.workspace = true
+opentelemetry_sdk.workspace = true
+opentelemetry.workspace = true
+tracing-opentelemetry.workspace = true
+
 
 #json序列化
 serde.workspace = true
@@ -40,10 +44,6 @@ uuid.workspace = true
 
 prost = "0.13.1"
 prost-types = "0.13.1"
-protobuf-json-mapping = '3.5.0'
-flate2 = "1.0.30"
-bzip2 = "0.4.4"
-tower = "0.4.13"
 
 pin-project-lite.workspace = true
 lazy_static.workspace = true
